Start UI migration to Blade templates
Using some cheap hosting providers, a standalone UI with Angular and connecting it to a Laravel as an api server can be very difficult to maintain and introduces strange bugs (especially with authentication).
If I move everything server-side, the UX for both me and end users can improve a lot (faster loading times, thanks to server-side caching and no need for multiple API calls to backend to check authentication).